feature: hydrationboundary, api 분리, 백엔드 api 기능 구현 완료
