+++ src/components/product/ProductReadMore.tsx
@@ -1,6 +1,5 @@
 'use client'
 import Image from 'next/image'
-// import react from 'React'
 import { useState } from 'react'
 import Button from '../common/Button/Button'
 export const ProductReadMore = ({ img }: { img: string }) => {
@@ -8,17 +7,16 @@ export const ProductReadMore = ({ img }: { img: string }) => {
 
   return (
     <div>
-      <div
-        className="w-full bg-white flex flex-col mt-4 justify-center items-center"
-        id="description"
-      >
+      <div className="w-full bg-white flex flex-col justify-center items-center " id="description">
         <div
-          className={`flex flex-col gap-4 transition-all duration-75 mt-4 ${expanded ? 'h-5000px' : 'max-h-[800px] overflow-hidden'}`}
+          className={`flex flex-col gap-4 transition-all duration-75 ${expanded ? 'h-5000px' : 'max-h-[800px] overflow-hidden'}`}
         >
           <Image width={420} height={420} src={img} alt="상푸 메인 이미지" />
         </div>
         <Button
-          className="bg-white text-stong border-strong border-[1px] rounded-none typo-button1"
+          className="rounded-none"
+          size="md"
+          variant={'outline'}
           onClick={() => setExpanded(!expanded)}
         >
           {expanded ? '상세설명 접기' : '상세설명 더보기'}
