+++ src/components/product/QuantitySelector.tsx
@@ -1,9 +1,21 @@
 'use client'
-// import { useState } from 'react'
+import { productStore } from '@/stores/productStore'
 import { NumbericField } from '../common/NumbericField/NumbericField'
+import { useEffect } from 'react'
+import formatKoreanWon from '@/util/formatKoreanWon'
 
-export const QuantitySelector = ({ productId }: { productId: number }) => {
-  // const [quantity, setQuantity] = useState<number>(1)
+export const QuantitySelector = ({
+  productId,
+  discountPrice,
+}: {
+  productId: number
+  discountPrice: number
+}) => {
+  const { quantity, setProductId } = productStore()
+
+  useEffect(() => {
+    if (setProductId && productId !== undefined) setProductId(productId)
+  }, [])
 
   return (
     <>
@@ -16,8 +28,10 @@ export const QuantitySelector = ({ productId }: { productId: number }) => {
       <div className="flex w-full justify-between items-center">
         <span className="typo-h3">총 금액</span>
         <div className="flex gap-2 items-center">
-          <span className="typo-body3 text-alternative">총 수량 1개</span>|
-          <span className="text-sale typo-h1">13,560원</span>
+          <span className="typo-body3 text-alternative">총 수량 {quantity}개</span>|
+          <span className="text-sale typo-h1">
+            {formatKoreanWon(discountPrice * quantity, false)}원
+          </span>
         </div>
       </div>
     </>
